requirements.txt
imageio==2.13.1
numpy==1.18.5
opencv_python==3.4.10.37
pandas==1.3.2
Pillow==9.2.0
scipy==1.7.1
torch==1.9.0+cu102
torchvision==0.10.0+cu102